mysite/requirements.txt
autopep8==1.6.0
Django==2.2
djangorestframework==3.13.1
Pillow==9.0.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
gunicorn